Struct tracing_filter::legacy::Filter

pub struct Filter { /* private fields */ }

A filter matching tracing-subscriber’s legacy EnvFilter format.

Implementations

impl Filter

pub fn parse(
spec: impl AsRef<str> + Into<String>
) -> (Filter, Option<ErrReport>)

Parse a filter from its string representation.

Filter compilation can produce warnings even when it succeeds, thus the nonstandard return type to provide an [ErrReport] on success.

impl Filter

pub fn new(spec: impl AsRef<str> + Into<String>) -> Self

Create a new filter, ignoring any invalid directives. It is highly recommended you use Self::parse instead, and display the warnings for ignored directives.

pub fn from_default_env() -> (Self, Option<ErrReport>)

Create a filter from the default RUST_LOG environment.

pub fn from_env(key: impl AsRef<OsStr>) -> (Self, Option<ErrReport>)

Create a filter from the environment.