Struct tracing_filter::FilterSubscriber

pub struct FilterSubscriber<F> { /* private fields */ }

A Subscriber which elevates a Filter from applying to a single subscriber to the entire layered subscribe stack.

Implementations

impl<F> FilterSubscriber<F>

pub fn new(filter: F) -> Self