Struct tracing_filter::legacy::Filter

pub struct Filter { /* private fields */ }

A filter matching tracing-subscriber’s legacy EnvFilter format.

Example

Use the RUST_LOG filter if it is set, but provide a fallback filter if the environment variable is not set.

let (filter, diagnostics) = Filter::from_default_env()
    .unwrap_or_else(|| Filter::parse("noisy_crate=warn,info"));

tracing_subscriber::registry()
    .with(filter.layer())
    .with(tracing_subscriber::fmt::layer())
    .init();

if let Some(diagnostics) = diagnostics {
    if let Some(error) = diagnostics.error(DiagnosticsTheme::default()) {
        error!("{error}");
    }
    if let Some(warn) = diagnostics.warn(DiagnosticsTheme::default()) {
        warn!("{warn}");
    }
}

Implementations

impl Filter

pub fn parse(spec: &str) -> (Filter, Option<Diagnostics<'_>>)

Parse a filter from its string representation.

Filter compilation can produce warnings even when it succeeds, thus the nonstandard return type to provide Diagnostics on success.

impl Filter

pub fn new(spec: &str) -> Self

Create a new filter, ignoring any invalid directives. It is highly recommended you use Self::parse instead, and display the warnings for ignored directives.

pub fn empty() -> Self

Create an empty filter (i.e. one that filters nothing out).

pub fn from_default_env() -> Option<(Self, Option<Diagnostics<'static>>)>

Create a filter from the default RUST_LOG environment.

pub fn from_env(
key: impl AsRef<OsStr>
) -> Option<(Self, Option<Diagnostics<'static>>)>

Create a filter from the environment.

Returns None if the environment variable is not set.

pub fn layer(self) -> FilterLayer<Self>

Lift this filter to a filter layer.